+++ packages/kad-dht/CHANGELOG.md
@@ -106,6 +106,44 @@
     * @libp2p/interface-compliance-tests bumped from ^5.3.0 to ^5.3.1
     * @libp2p/peer-store bumped from ^10.0.9 to ^10.0.10
 
+## [16.2.6](https://github.com/libp2p/js-libp2p/compare/kad-dht-v16.2.5...kad-dht-v16.2.6) (2026-05-13)
+
+
+### Bug Fixes
+
+* **kad-dht:** only echo PUT_VALUE response on success ([#3502](https://github.com/libp2p/js-libp2p/issues/3502)) ([5692f3f](https://github.com/libp2p/js-libp2p/commit/5692f3f168049cc658ada36e8251b445683f5806))
+
+## [16.2.5](https://github.com/libp2p/js-libp2p/compare/kad-dht-v16.2.4...kad-dht-v16.2.5) (2026-05-09)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/crypto bumped from ^5.1.17 to ^5.1.18
+    * @libp2p/interface-internal bumped from ^3.1.3 to ^3.1.4
+    * @libp2p/peer-collections bumped from ^7.0.18 to ^7.0.19
+    * @libp2p/peer-id bumped from ^6.0.8 to ^6.0.9
+    * @libp2p/ping bumped from ^3.1.3 to ^3.1.4
+    * @libp2p/utils bumped from ^7.1.0 to ^7.2.0
+  * devDependencies
+    * @libp2p/logger bumped from ^6.2.6 to ^6.2.7
+    * @libp2p/peer-store bumped from ^12.0.18 to ^12.0.19
+
+## [16.2.4](https://github.com/libp2p/js-libp2p/compare/kad-dht-v16.2.3...kad-dht-v16.2.4) (2026-04-25)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface-internal bumped from ^3.1.2 to ^3.1.3
+    * @libp2p/peer-collections bumped from ^7.0.17 to ^7.0.18
+    * @libp2p/ping bumped from ^3.1.2 to ^3.1.3
+    * @libp2p/utils bumped from ^7.0.17 to ^7.1.0
+  * devDependencies
+    * @libp2p/peer-store bumped from ^12.0.17 to ^12.0.18
+
 ## [16.2.3](https://github.com/libp2p/js-libp2p/compare/kad-dht-v16.2.2...kad-dht-v16.2.3) (2026-04-20)