+++ packages/peer-discovery-bootstrap/CHANGELOG.md
@@ -83,6 +83,39 @@
   * devDependencies
     * @libp2p/interface-compliance-tests bumped from ^5.3.0 to ^5.3.1
 
+## [12.0.21](https://github.com/libp2p/js-libp2p/compare/bootstrap-v12.0.20...bootstrap-v12.0.21) (2026-05-09)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface-internal bumped from ^3.1.3 to ^3.1.4
+    * @libp2p/peer-id bumped from ^6.0.8 to ^6.0.9
+  * devDependencies
+    * @libp2p/interface-compliance-tests bumped from ^7.0.20 to ^7.0.21
+    * @libp2p/logger bumped from ^6.2.6 to ^6.2.7
+
+## [12.0.20](https://github.com/libp2p/js-libp2p/compare/bootstrap-v12.0.19...bootstrap-v12.0.20) (2026-05-04)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @libp2p/interface-compliance-tests bumped from ^7.0.19 to ^7.0.20
+
+## [12.0.19](https://github.com/libp2p/js-libp2p/compare/bootstrap-v12.0.18...bootstrap-v12.0.19) (2026-04-25)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface-internal bumped from ^3.1.2 to ^3.1.3
+  * devDependencies
+    * @libp2p/interface-compliance-tests bumped from ^7.0.18 to ^7.0.19
+
 ## [12.0.18](https://github.com/libp2p/js-libp2p/compare/bootstrap-v12.0.17...bootstrap-v12.0.18) (2026-04-16)