+++ packages/pnet/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## [3.0.20](https://github.com/libp2p/js-libp2p/compare/pnet-v3.0.19...pnet-v3.0.20) (2026-05-09)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/crypto bumped from ^5.1.17 to ^5.1.18
+    * @libp2p/utils bumped from ^7.1.0 to ^7.2.0
+
+## [3.0.19](https://github.com/libp2p/js-libp2p/compare/pnet-v3.0.18...pnet-v3.0.19) (2026-04-25)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/utils bumped from ^7.0.17 to ^7.1.0
+
 ## [3.0.18](https://github.com/libp2p/js-libp2p/compare/pnet-v3.0.17...pnet-v3.0.18) (2026-04-16)